generated_requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
pyyaml>=6.0
requests>=2.28.0
apache-airflow>=2.5.0
psycopg2-binary>=2.9.0